Transformer_fix

+++ Transformers/helper.py
@@ -10,9 +10,9 @@ import torch.utils.data as Data
 class MyDataSet(Data.Dataset):
     def __init__(self, enc_inputs, dec_inputs, dec_outputs):
         super(MyDataSet, self).__init__()
-        self.enc_inputs = torch.tensor(enc_inputs)
-        self.dec_inputs = torch.tensor(dec_inputs)
-        self.dec_outputs = torch.tensor(dec_outputs)
+        self.enc_inputs = torch.tensor(enc_inputs,dtype=torch.int32)
+        self.dec_inputs = torch.tensor(dec_inputs,dtype=torch.int32)
+        self.dec_outputs = torch.tensor(dec_outputs,dtype=torch.int32)
 
     def __len__(self):
         return self.enc_inputs.shape[0]
